Style

+++ src/controllers/videoController.js
@@ -2,7 +2,7 @@ import Video from "../models/Video";
 import User from "../models/User";
 
 export const home = async (req, res) => {
-    const videos = await Video.find({});
+    const videos = await Video.find({}).populate("owner");
     return res.render("home", { pageTitle: "Home", videos});
 };
     
@@ -17,20 +17,28 @@ export const watch = async (req, res) => {
 };
 export const getEdit = async (req, res) => {
     const { id } = req.params;
+    const { user: { _id } } = req.session;
     const video = await Video.findById(id);
     if (!video) {
         return res.status(404).render("404", {pageTitle: "Video not found."});
+    }
+    if (String(video.owner) !== _id) {
+        return res.status(403).redirect("/");
     }
     
     return res.render("edit", { pageTitle: `Edit ${video.title}`, video});
 };
 
 export const postEdit = async (req, res) => {
+    const { user: { _id } } = req.session;
     const { id } = req.params;
     const { title, description, hashtags } = req.body;
     const video = await Video.exists({_id: id});
     if (!video) {
         return res.status(404).render("404", {pageTitle: "Video not found."});
+    }
+    if (String(video.owner) !== _id) {
+        return res.status(403).redirect("/");
     }
     await Video.findByIdAndUpdate(id, {
         title, 
@@ -49,13 +57,16 @@ export const postUpload = async (req, res) => {
     const {path: fileUrl} = req.file;
     const { title, description, hashtags } = req.body;
     try {
-        await Video.create({
+        const newVideo = await Video.create({
             title,
             description,
             hashtags: Video.formatHashtags(hashtags),
             fileUrl,
             owner: _id,
         });
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        await user.save()
         return res.redirect("/");
     } catch (error) {
         console.log(error);
@@ -66,7 +77,15 @@ export const postUpload = async (req, res) => {
 };
 
 export const deleteVideo = async (req, res) => {
+    const { user: { _id } } = req.session;
     const { id } = req.params;
+    const video = await Video.findById(id);
+    if (!video) {
+        return res.status(404).render("404", {pageTitle: "Video not found."});
+    }
+    if (String(video.owner) !== _id) {
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
     return res.redirect('/');
 };
@@ -79,7 +98,7 @@ export const search = async (req, res) => {
             title: {
                 $regex: new RegExp(`^${keyword}`, "i"),
             },
-        });
+        }).populate("owner");
     }
     return res.render("search", {pageTitle: "Search", videos});
 };
